create storage module
