+++ index.html
@@ -74,6 +74,6 @@
       <span>Copyright</span>
     </footer>
 
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
